+++ D3DEngine/Source/Core/Timer.cpp
@@ -1,7 +1,7 @@
 // Timer.cc
 #include "D3DEngine/Core/Timer.h"
 
-double Timer::_precision = 0.0;
+float Timer::_precision = 0.0;
 
 void Timer::InitTimer() {
 	LARGE_INTEGER freq;
@@ -12,8 +12,8 @@ void Timer::InitTimer() {
 	_precision = 1.0 / static_cast<double>(freq.QuadPart);
 }
 
-double Timer::GetTick() {
+float Timer::GetTick() {
 	LARGE_INTEGER counter;
 	QueryPerformanceCounter(&counter);
-	return static_cast<double>(counter.QuadPart) * _precision;
+	return static_cast<float>(counter.QuadPart) * _precision;
 }
